pyproject.toml
[project]
name = "basango"
version = "0.1.0"
description = "Basango : Web Scrapper for DRC News"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "rq>=2.5.0",
    "typer>=0.16.1",
    "uv-build>=0.8.12,<0.9.0",
    "pyyaml>=6.0.2",
    "httpx>=0.27.2",
    "trafilatura>=1.7.0",
    "selectolax>=0.3.20",
    "markdownify>=0.13.1",
    "readability-lxml>=0.8.1",
]

[dependency-groups]
dev = [
    "bandit>=1.8.6",
    "pyright>=1.1.404",
    "pytest>=8.4.1",
    "ruff>=0.12.9",
]

[project.scripts]
basango = "basango:main"

[build-system]
requires = ["uv_build>=0.8.12,<0.9.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
testpaths = ["tests"]
